Allow dryrun with a non-existent keyfile in config
tarsnap.conf.sample comes with a default location for the keyfile, but there
is no keyfile there until the user runs tarsnap-keygen, so

    $ tarsnap --dry-run -c ~/backup

fails.  Users could avoid this problem by specifying --no-default-config on
the command-line, but it would be nicer if tarsnap itself could forgive having
a non-existent keyfile specified in the config file (but only the config file,
not the command-line!).