Multitape layer and cookie for printing stats for archive creation
Unfortunately writetape_close() is called by write_close(), which only has
access to the TAPE_W cookie.  As a result, passing ${csv_filename} to this
function was much messier than the other "printing statitics" which were
implemented in tar/glue/tape.c.

This adds "csv_filename" to the function call, which has been set to NULL in
order to retain the original program behaviour for now.

No functionality has been connected to the main executable.